Small Updates

+++ cpp/p002.cpp
@@ -14,14 +14,14 @@ int compute(const int limit) {
 	const double phi = (1 + sqrt(5)) / 2;
 	
 	/*
-	 * - nth fibonacci term formula: nth = [Phi^n – (phi)^n] / Sqrt[5]
+	 * - nth fibonacci term formula: nth = (phi^n – (1 - phi)^n) / sqrt[5]
 	 * - Every third term in a fibonacci sequence is even.
 	 */
 	
 	int sum = 0;
 	int term = 3;
 	while (sum < (limit / 3)) {
-		sum += static_cast<int>((pow(phi, term) - pow((1 - phi), term)) / sqrt(5));
+		sum += (pow(phi, term) - pow((1 - phi), term)) / sqrt(5);
 		term += 3;
 	}
